+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 4.1.0
 
 - Ability to resolve `.zil` domains from Polygon and Ethereum networks
+- Change default blockchain provider from Infura to Alchemy
 ### Breaking changes
 - ZNS support moved under UNS nameserice
 - Remove separate ZNS config